docs: critical docstring quality pass across the public API
Audit + fix every public class/method/function docstring for accuracy,
clarity, and consistency with the implementation (beyond the structural
house-style lint).

Fixes (doc-vs-code drift, copy-paste, contradictions):
- Defaults/types aligned to signatures (figsize, component_y, len_max,
  n_cols, max_cor, pc_x/pc_y, fontsize_annotations, cbar_xywh, ...).
- encode_integer Returns ("integer", not "one-hot"); AAclust comp_centers/
  comp_medoids Returns shapes -> (n_clusters, n_features); CPP.run_num made
  self-contained (repeat shared params, no "See run" lump); AAWindowSampler/
  scan_motif Returns df_seq -> df_seq_out; comp_seq_sim split return branches.
- Module-docstring typos "Mote Carlo" -> "Monte Carlo"; ShapModel grammar;
  TreeModel false "requires SHAP" removed; dPULearnPlot summary rewritten.
- Documented previously-undocumented params (plddt_disorder_threshold,
  CPP.run/run_num return_stats).

Incidental code-correctness fixes:
- dPULearnPlot.__int__ -> __init__ (constructor never ran).
- NumericalFeature.filter_correlation validation name="min_cor" -> "max_cor".
- random_state annotation Optional[str] -> Optional[int] (CPP, AAclustPlot,
  dPULearn); ShapModel.exp_value Optional[int] -> Optional[float].

Tooling (docstring-consistency checker):
- New XREF-UNRESOLVED (cross-reference integrity, package-wide registry).
- New RAISES-UNDOCUMENTED (advisory); CITATION-UNDEFINED already added.
- Stub skipping + Defects/Advisory split so "0 defects" = clean.
- Style guide: self-contained-params rule, cross-ref rule, citation rules,
  build-clean gate; conformance table updated.

Verification: check_docstrings 0 defects; pytest tests/unit 2950 passed.

+++ aaanalysis/plotting/_plot_gcfs.py
@@ -20,7 +20,7 @@ def plot_gcfs(option: str = 'font.size') -> int:
 
     Returns
     -------
-    option_value : int
+    option_value : float
         Numerical value for selected option.
 
     See Also
